chore: 불필요한 VO파일 삭제 -  Dto파일에서 url 생성 로직 추가
