+++ src/main/java/com/flab/mars/api/controller/StockController.java
@@ -5,7 +5,6 @@ import com.flab.mars.api.dto.response.ResultAPIDto;
 import com.flab.mars.api.dto.response.StockFluctuationDto;
 import com.flab.mars.domain.service.StockService;
 import com.flab.mars.domain.vo.StockPrice;
-import com.flab.mars.domain.vo.request.StockFluctuationRequestVO;
 import com.flab.mars.domain.vo.response.StockFluctuationResponseVO;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
@@ -33,18 +32,7 @@ public class StockController {
     @GetMapping("/domestic-stock/ranking/fluctuation")
     public ResponseEntity<ResultAPIDto<StockFluctuationDto>> getFluctuationRanking(@ModelAttribute @Valid StockFluctuationRequestDto request, HttpSession session) {
 
-        StockFluctuationRequestVO vo = StockFluctuationRequestVO.builder()
-                .fidInputIscd(request.getFidInputIscd())  // 기본값: "0001"
-                .fidRankSortClsCode(request.getFidRankSortClsCode())  // 기본값: "0"
-                .fidInputCnt1(request.getFidInputCnt1())  // 기본값: "0"
-                .fidPrcClsCode(request.getFidPrcClsCode())  // 기본값: "0"
-                .fidInputPrice1(request.getFidInputPrice1())  // 기본값: ""
-                .fidInputPrice2(request.getFidInputPrice2())  // 기본값: ""
-                .fidVolCnt(request.getFidVolCnt())  // 기본값: ""
-                .build();
-
-        String uri = vo.generateFluctuationRankingUri();
-
+        String uri = request.generateFluctuationRankingUri();
         StockFluctuationResponseVO response = stockService.getFluctuationRanking(uri, session);
         StockFluctuationDto dto = StockFluctuationDto.toDTO(response);
